move desktop tests
