+++ td.desktop/.jshintignore
@@ -1,1 +1,3 @@
+test/
 app/threatdragon.js
+core/templates.js
