+++ td.desktop/test/core/diagrams/diagram_spec.js
