+++ td.desktop/test/core/diagrams/diagramdirectives_spec.js
