+++ td.desktop/test/core/diagrams/elementpropertydirectives_spec.js
