+++ td.desktop/test/layout/shell_spec.js
