+++ td.site/.jshintignore
@@ -1,2 +1,1 @@
 test/
-src/app/core/templates.js
